Centralize clima params def and construction
