+++ examples/hybrid/utilities.jl
@@ -1,0 +1,19 @@
+function time_to_seconds(s::String)
+    factor = Dict(
+        "secs" => 1,
+        "mins" => 60,
+        "hours" => 60 * 60,
+        "days" => 60 * 60 * 24,
+    )
+    s == "Inf" && return Inf
+    if count(occursin.(keys(factor), Ref(s))) != 1
+        error(
+            "Bad format for flag $s. Examples: [`10secs`, `20mins`, `30hours`, `40days`]",
+        )
+    end
+    for match in keys(factor)
+        occursin(match, s) || continue
+        return parse(Float64, first(split(s, match))) * factor[match]
+    end
+    error("Uncaught case in computing time from given string.")
+end
